% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chroot.R
\name{chroot}
\alias{chroot}
\title{Change Root Dir}
\usage{
chroot(path = getwd())
}
\arguments{
\item{path}{directory of the new root}
}
\description{
Changes the root directory of the calling process to that specified in path.
This directory will be used for pathnames beginning with \code{/}.
\strong{Only a privileged process (i.e. sudo) may call \code{chroot()}}.
}
\details{
This call changes an ingredient in the pathname resolution process
and does nothing else.  In particular, it is not intended to be used
for any kind of security purpose, neither to fully sandbox a process
nor to restrict filesystem system calls.
}
\references{
\href{http://man7.org/linux/man-pages/man2/chroot.2.html}{CHROOT(2)}
}
