% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userinfo.R
\name{userinfo}
\alias{userinfo}
\alias{user_info}
\alias{group_info}
\title{User / Group Info}
\usage{
user_info(uid = getuid())

group_info(gid = getgid())
}
\arguments{
\item{uid}{user ID (integer) or name (string)}

\item{gid}{group ID (integer) or name (string)}
}
\description{
Lookup a user or group info via user uid/name or group gid/name.
}
\examples{
# Get info current user
user_info()
group_info()
}
\references{
\href{https://man7.org/linux/man-pages/man3/getpwnam.3.html}{GETPWNAM(3)}
\href{https://man7.org/linux/man-pages/man3/getgrnam.3.html}{GETGRNAM(3)}
}
