% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getuid}
\alias{getuid}
\alias{getgid}
\alias{getpid}
\alias{getppid}
\alias{getpgid}
\alias{getprio}
\alias{setuid}
\alias{setgid}
\alias{setpgid}
\alias{setprio}
\alias{kill}
\title{Process Info}
\usage{
getuid()

getgid()

getpid()

getppid()

getpgid()

getprio()

setuid(uid)

setgid(gid)

setpgid(pgid = 0)

setprio(prio)

kill(pid, signal = SIGTERM)
}
\arguments{
\item{uid}{User ID from \code{/etc/passwd}.}

\item{gid}{Group ID from \code{/etc/group}.}

\item{pgid}{Process Group ID. Default \code{0} sets pgid to the current pid.}

\item{prio}{Priority level}

\item{pid}{process ID (integer)}

\item{signal}{a signal number (integer), defaults to \link[tools:SIGTERM]{tools::SIGTERM}.}
}
\description{
Get or set attributes of the current process.
}
\details{
Acronyms stand for:
\itemize{
\item \code{pid} Process ID
\item \code{ppid} Parent-Process ID
\item \code{pgid} Process-Group ID
\item \code{uid} User ID
\item \code{gid} Group ID
\item \code{prio} Priority level
}

An unprivileged (non-root) process cannot change it's \code{uid} and only lower
process priority (higher value).
}
