% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{getuid}
\alias{getuid}
\alias{getgid}
\alias{getpid}
\alias{getppid}
\alias{getpgid}
\alias{setuid}
\alias{setgid}
\alias{setpgid}
\alias{kill}
\title{Process Info}
\usage{
getuid()

getgid()

getpid()

getppid()

getpgid()

setuid(uid)

setgid(gid)

setpgid(pgid = 0)

kill(pid, signal = SIGTERM)
}
\arguments{
\item{uid}{User ID from \code{/etc/passwd}.}

\item{gid}{Group ID from \code{/etc/group}.}

\item{pgid}{Process Group ID. Usually must be the same as current group}

\item{pid}{process ID (integer)}

\item{signal}{a signal number (integer), defaults to \link[tools:SIGTERM]{tools::SIGTERM}.}
}
\description{
Get or set attributes of the current process. Acronyms stand for:
}
\details{
\itemize{
\item \code{pid} Process ID
\item \code{ppid} Parent-Process ID
\item \code{pgid} Process-Group ID
\item \code{uid} User ID
\item \code{gid} Group ID
}
}
