% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/userinfo.R
\name{user_info}
\alias{user_info}
\alias{group_info}
\alias{me}
\title{Lookup UID/GRP info}
\usage{
user_info(uid = me())

group_info(gid = me())

me()
}
\arguments{
\item{uid}{user ID (integer) or name (string)}

\item{gid}{group ID (integer) or name (string)}
}
\description{
Lookup user or group info via UID/GID or user/group name
}
\examples{
# Get info about root
user_info(0)
group_info(0)
}
